deploy/client/pi-requirements.txt
Adafruit-BMP==1.5.2
Adafruit-DHT==1.3.2
Adafruit-GPIO==1.0.3
Adafruit-PureIO==0.2.1
altgraph==0.15
certifi==2018.1.18
chardet==3.0.4
future==0.16.0
gps3==0.33.3
idna==2.6
macholib==1.9
numpy==1.14.2
pathlib==1.0.1
pefile==2017.11.5
Pillow==5.0.0
pkg-resources==0.0.0
PyInstaller==3.3.1
PyYAML==3.12
requests==2.18.4
RTIMULib==7.2.1
sense-emu==1.0
sense-hat==2.2.0
spidev==3.2
urllib3==1.22
